oracle-monitor/agents/requirements.txt
requests==2.31.0
pyyaml==6.0.1
kafka-python==2.0.2
